A bunch of coersions, fmt changes, type long to int (which eliminated a long-standing warning from MSVC in 32-bit mode).  I don't know why some of these warnings just showed up now (maybe Michael's cygwin compiler invocation change?), but good to have them.
git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@8383 756869e0-9e47-4bc5-b0a8-847305dc5e1c

+++ XSB/emu/system_xsb.c
@@ -938,12 +938,12 @@ static char *get_next_command_argument(char **buffptr, char **cmdlineprt)
   char *returnptr = *buffptr;
 
   /* skip white space */
-  while (isspace(**cmdlineprt))
+  while (isspace((int)**cmdlineprt))
     (*cmdlineprt)++;
 
   /* loop as long as not end of cmd line or until the next command argument has
      been found and extracted */
-  while ((!isspace(**cmdlineprt) || quoted) && **cmdlineprt != '\0') {
+  while ((!isspace((int)**cmdlineprt) || quoted) && **cmdlineprt != '\0') {
     if (escaped) {
       switch (**cmdlineprt) {
       case 'b': **buffptr='\b'; break;
